+++ Altruist/Gaming/Main/Core/General/Engine/Engine.cs
@@ -457,7 +457,7 @@ public class AltruistEngine : IAltruistEngine
         long previousTicks = FrameTime.NowTicks;
         // 15 FPS = ~66.6ms
         // TODO: it will have a big buggy bug if engine rate is not Hz!!
-        long tickLength = (long)(Stopwatch.Frequency / _engineRate.Value);
+        long tickLength = Stopwatch.Frequency / _engineRate.Value;
 
         while (!_cancellationTokenSource.Token.IsCancellationRequested)
         {
